[REFACTOR] single choice fragment,viewmodel 리팩토링 진행
