+++ app/src/main/java/com/omoolen/omooroid/onboarding/fragments/one/OneOnboardViewModel.kt
@@ -8,6 +8,8 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
 import androidx.recyclerview.widget.RecyclerView
 import com.omoolen.omooroid.R
+import com.omoolen.omooroid.onboarding.fragments.one.recycle.age.AgeAdapter
+import com.omoolen.omooroid.onboarding.fragments.one.recycle.age.AgeInfo
 import com.omoolen.omooroid.onboarding.fragments.one.recycle.gender.GenderAdapter
 import com.omoolen.omooroid.onboarding.fragments.one.recycle.gender.GenderInfo
 import com.omoolen.omooroid.util.ListLiveData
@@ -16,11 +18,12 @@ class OneOnboardViewModel(application: Application) : AndroidViewModel(applicati
     @SuppressLint("StaticFieldLeak")
     private val mContext = getApplication<Application>().applicationContext
     private lateinit var genderAdapter: GenderAdapter
+    private lateinit var ageAdapter: AgeAdapter
+
     //var genderArr = ListLiveData <Boolean>()
-    var ageArr = ListLiveData <Boolean>()
+    //var ageArr = ListLiveData <Boolean>()
 
     fun setGenderAdapter() :GenderAdapter{
-        //TODO : 초기화 ->나중에 refactor하기
         genderAdapter = GenderAdapter()
         genderAdapter.genderList.addAll(
             listOf<GenderInfo>(
@@ -28,7 +31,28 @@ class OneOnboardViewModel(application: Application) : AndroidViewModel(applicati
                 GenderInfo(resourceId = R.drawable.onboard_female_selector,name = "남성")
             )
         )
+        genderAdapter.setItemClickListener(object : GenderAdapter.OnItemClickListener {
+            override fun onClick(v: View, position: Int) {
+            }
+        })
         return genderAdapter
+    }
+
+    fun setAgeAdapter() :AgeAdapter{
+        ageAdapter = AgeAdapter()
+        ageAdapter.ageList.addAll(
+            listOf<AgeInfo>(
+                AgeInfo(age = "10대"),
+                AgeInfo(age = "20대"),
+                AgeInfo(age = "30대"),
+                AgeInfo(age = "40대 이상")
+            )
+        )
+        ageAdapter.setItemClickListener(object : AgeAdapter.OnItemClickListener {
+            override fun onClick(v: View, position: Int) {
+            }
+        })
+        return ageAdapter
     }
 
     fun singleChoice(rv: RecyclerView, e: MotionEvent){
